Fix operator precedence
Ensure that unary operators have higher precedence than the hat
operator.  Ensure the correct operands are used for unary and hat so
that parser doesn't error out with unbalanced.  Special case hatOperand
to be either unaryOperation or unaryOperand.  This allows any expression
to degrade to just a unaryOperand, which is the old expression rules of
atom, CHAR_TOKEN, NUMBER, charListHeredoc, quote, sigil, and
stringHeredoc, so that operatorless operands don't all become bad
characters until all operators are finished.