Addition (+/-) operations
Break out '+' and '-' from UNARY_OPERATOR and put them in DUAL_OPERATOR.
It's called dual because '+' and '-' can be parsed as either a unary or
a binary operator.  From experiments with Code.string_to_quoted/1, it
appears that dual operators are binary when there is no EOL between the
first argument and the operator and unary otherwise.