Disallow <quote><space><colon>
When encountering a `QUOTE_PROMOTER` (`"` ir `'`) put the return state as
`KEYWORD_PAIR_MAYBE` instead of the current state.  The
`KEYWORD_PAIR_MAYBE` state checked for `{COLON} / {SPACE}`.  If it is
found, it is lexed as `KEYWORD_PAIR_COLON` and the lexer returns to the
state where `QUOTE_PROMOTER` was detected. If `{COLON} / {SPACE}` is not found,
then the token is handled in the state where `QUOTE_PROMOTER` was detected.