Pass all potential keywordKeys through KEYWORD_PAIR_MAYBE
Instead of detecting keywordKeys for operators, ALIASes, and IDENTIFIERs
using the `/ {COLON}{SPACE}` trailing syntax and lexing them as
`KEYWORD_KEY_LITERAL` followed by `COLON`, have all operators, `ALIAS`,
and `IDENTIFIER` go the the `KEYWORD_PAIR_MAYBE` state, which will parse
`{COLON} / {SPACE}` as `KEYWORD_PAIR_COLON` that was used with `quote`.
This leaves the `COLON` token type available to be parsed for a
correctable error when there is a space between the `ALIAS`,
`IDENTIFIER`, operator, or `quote` and the `COLON`.