Binary and unknown base numbers with invalid digits
Instead of immediately ending a binary number when a a digit besides 0
or 1 is found, lex runs of 0 and 1 as VALID_BINARY_DIGITS and all other
digits in [A-Za-z2-9]+ as INVALID_BINARY_DIGITS, which allows for
parsing the invalid binary number in the pygment example without
breaking the parsing of the rest of the file.  unknownBaseNumber extends
this error handling to the base-specifier, so if an alphabetical
character besides b, x, or o is given after 0, then that is parsed as an
unknownBaseNumber.  All alphanumeric characters are considered
INVALID_UNKNOWN_BASE_DIGITS.