Allow invalid digits in decimal integers
Allow invalid digits, [A-Fa-f]+, in decimal integers.  [0-9]+ is lexed
as VALID_DECIMAL_DIGITS, while [A-Fa-f]+ is lexed as
INVALID_DECIMAL_DIGITS.  '_' is lexed as DECIMAL_SEPARATOR and can go
between valid or invalid decimal digits.  The only restrictions on
decimalNumber is that it must start with a VALID_DECIMAL_DIGITS and if
the first digit is a 0 the second can't be an invalid decimal digit or
that will trigger BASE_INTEGER_BASE as the invalid decimal digit, a
letter, will be interrupted as either a known or unknown base specifier.