Fix dot_alias and dot_identifier translations
matchedExpressionDotOperation allowed any matchedExpression for the
right operand when it should have only allowed ALIAS or IDENTIFIER.
Transliterating dot_alias and dot_identifier leads to left-recursive
that is unterminable because the parser generator can only handle
left-recursion in expressions that are direct children, so dot_alias is
split into matchedExpressionDotAlias and ALIAS in maxExpression.
maxExpression along with accessExpression had matchedExpression- added
to their name to signify they are only meant for use as descendants of
matchedExpression and not direct transliterations anymore.
dot_identifier is split into matchedExpressionDotIdentifier and
identifierExpression.  There can't be an rule representing
no_parens_one_expression as it will lead to an indirect descendent
starting with matchedExpression, which leads to non-terminating
left-recursion.