matched_expr -> at_op_eol no_parens_expr
Add `matchedExpressionNoParenthesesExpressionAtOperation` to
`matchedExpression` before `matchedExpressionAtOperation` because
`noParenthesesExpression` matches a superstring of `matchedExpression`.
`matchedExpressionNoParenthesesExpressionAtOperation` is prefixed
`matcheExpression` because it needs to extend `matchedExpression` to
prevent unresolved left-recursion in the generated parser.  The rule can
still be written in the obvious way as
`matchedExpressionNoParenthesesExpressionAtOperation ::= AT_OPERATOR
EOL* noParenthesesExpression`, it's the extension of matchedExpression,
which is crucial.  This approach should be generalized back to
`matchedExpressionCaptureOperation` to split it into a number,
noParenthesesExpression and matchedExpression version.
`matchedExpressioNCaptureOperation` only works now as written because it
is the first ordered choice in matchedExpression.  The approach should
also be used for `matched_expr -> unary_op_eol no_parens_expr`.