matched_expr -> matched_expr no_parens_op_expr
When translating `matched_expr -> matched_expr no_parens_op_expr` and
using `Code.string_to_quoted/1` to check for correct precedence, I
realized that `no_parens_op_expr` should only be allowed as the final,
right most operand of any nesting of operations.  It is not possible to
guarantee noParenthesesExpression is only the right-most operand due to
the recursive nature of matchedExpression.  Therefore, all the
operations with noParenthesesExpression as the right operand need to be
in `noParenthesesExpression` instead of in matchedExpression.  I also
started naming the operations in the order of the operands and operator
instead of <operand><operator>Operation as I was doing before.  This
naming scheme needs to be applied to matchedExpression too.