Restructure roots to allow noParenthesesExpression in tail
The Pratt parsing triggered by matchedExpression had built in error
handling pinned to the operator, so if the right-hand operand, such as
an identifier NOT followed by callArgumentsNoParenthesesManyStrict, does
not match because it's actually is `identifier
callArgumentsNotParenthesesManyStrict` for noParenthesesExpression, then
the overall matchedExpression would still match and recover.  This would
man that noParenthesesExpression would never actually match.  Since I
couldn't figure out how make matchedExpression as a left operand while
matchedExpression Or noParenthesesExpression on the right operand, I had
to abandon the Pratt parsing system and use a normal Left-to-Right
Left-most deriviation (LL) grammar.  There does appear to be performance
issues with nested interpolation.