matched + tail -> matched + expression
Instead of having completely independent matchedExpression and
tailExpression hierarchies, have an expression hierarchy that is
composed of repeated matchedExpression operations followed by a
expression operation or an expression operation when there is no
matchedExpressions.  This pattern allows for less backtracking at
matchedExpressions match as quickly as possible or fallback to just
expressions and finally tailExpression.  This change has eliminated the
performance programs with nested interpolation tests.  While reviewing
the testData changes, I realized that abandonning the generated Pratt
parser from matchedExpression meant that I lost prefix operation
nesting, so I had to manually at to each Operation the operation to
match lower precedence prefix operations.  Importantly, these prefix
operation must have the prefixOperator, so they are called
PrefixOperation while when the prefixOperator is optional, is just an
Operation as is the case with binay operations for infix operations.