Always use captureExpressionPrefixOperation
Remove captureMatchedExpressionPrefixOperation and use
captureExpressionPrefixOperation in its place.  Because capture is a
prefix operator, it consumes all tokens to the right and so it can
safely always consume expressions and not just matchedExpressions
because any matchedExpression + expression to the right should be
treated as a single argument to the capture.  This change fixes issues
with the captureAliasHatCaptureAliasHatCaptureTailExpression tests.
Still need to determine if other prefix operators should always use
expressions and not matched expressions also or if this behavior for
capture is due to be both a prefix operator and having the highest
precedence.