Group arguments with receiver in dot expression
Instead of having a rule similar to `dot_identifier`, have each type of
function call be a `left` operation that groups the final identifier and
the arguments together in one call.  This will eliminate the need to
have `!` negative look-ahead, but there is still a need for `&` positive
look-ahead so that `matchedDotOperation` doesn't accidentally consume
the final IDENTIFIER, FALSE, NIL, TRUE, binaryString, or listString and
instead allows it to be consumed by
noParenthesesNoArgumentedQualifiedCall or qualifiedAlias.