Collapse right and left sides together
Use `operand operation*` with `left operation` to limit backtracking.
Allow noParenthesesManyArgumentsCall into matchedExpression by having an
operation level for matchedCallOperation that requires an
variable or accessExpression or below on the right and
noParenthesesManyArgumentsStrict on the left.  There is no actual
operator in this operation.  The space could be thought of as a virtual
operator.