Make accessExpression public
accessExpression must be public so that it can extend matchedExpression
instead of have its descendents extend matchedExpression.  It is
important that the accessExpression descendents DO NOT extend
matchedExpression as those descendants can't be used in negative look-aheads
as !<matchedExpression descendant> is converted to !matchedExpression in
the parser when the descendants extend matchedExpression.