Split up matchedWhenOperation to fix operator precedence
Having `(noParenthesesKeywords | matchedExpression)` as the right
operand in matchedWhenOperation caused it to be a POSTFIX operation
instead of a BINARY operation, which caused the operator precedence
between matchedInMatchOperation and matchedWhenOperation to be wrong
when there were matchedWhenOperations are both sides of
matchedInMatchOperation.  To fix the problem, matchedWhenOperation has
only matchedExpression for both operands, making it BINARY again and
there is a separate matchedWhenNoParenthesesKeywordsOperation which is a
POSTFIX operation and places before matchedWhenOperation in
matchedExpression as the Pratt parser's error handling in
matchedExpression will make matchedWhenOperation's right
matchedExpression always match, but with an internal PsiErrorElement and
therefore block matchedWhenNoParenthesesKeywordsOperation from ever
matching.