Differentiate between escape sequences in sigils and quotes
`\x` causes an ArgumentError in quotes (charListLine, charListHeredoc,
stringLine, stringHeredoc), but not in sigils in native Elixir because
quoting sigils doesn't evaluate the contents of the Elixir string passed
to the sigil function.  This means that `\x` should only be marked as an
error in intellij-elixir inside quotes and not sigils, so the body and
escape sequences need to be differentiated between quotes and sigils.