Bracket Expression with proper nesting of calls
To properly distinguish between `foo<space>[]` and `foo[]`, break up
calls into (1) no parentheses calls, (2) parentheses calls, (3) no
argument calls.  Calls are already broken up by (A) qualified and
(B) unqualified.  These leads to the combinations (A1)
matchedQualifiedNoParenthesesCall, (B1) matchedUnqualifedNoParenthesesCall,
(A2) matchedQualifiedParenthesesCall, (A3) matchedQualifiedNoArgumentsCall,
(B2) matchedUnqualifiedParenthesesCall, and (B3) variable.

Separating (1) and (2) allowed matchedBracketOperation to be placed
after (1) and before (2).  Being after (1) makes sense because brackets
immediately after a no parentheses call would attach to one of the
arguments of the call and not the call itself.

matchedAtNonNumericOperation needs to be before any unqualified call,
but this included matchedUnqualifiedNoParenthesesCall, which needed to
be before matchedBracketOperation and therefore
matchedAtNonNumericOperation.  Since these two constraints are
inconsistent, matchedAtUnqualifiedNoParenthesesCall was added to handle
the need for matchedAt*Operation to wrap
matchedUnqualifedNoParenthesesCall.

During the testing of matchedAtUnqualifiedNoParenthesesCall, debugging
the parsing revealed that if the parser generated a PREFIX operation for
the unqualified calls, that was incorrect.  The operations were PREFIX
because they used variable and variable extends matchedExpression.  The
actual implementation of the parser passing the `g` of variable, so it
or any lower priority pattern could match.  This was not the intent when
using variable, so variable was replaced with the IDENTIFIER token,
which made the rules ATOM instead of PREFIX.