fn -> <expression>+ end
Use block quoting pattern from quote(ELixirFile) in
quote(ElixirStabBody) as block properly returns `NIL` when no children
are present as quote(ElixirStabBody) previously did and handles multiple
children correctly.  The only difference is that when there is a single
unary operation, a single element block is used instead of the quoted
element.  This was the previous behavior of quote(ElixirStabBody).