Replace stabParenthesesManyArguments with parenthesesArguments
It turns out that `fn (...) -> ... end` allows a single argument and
not just many arguments in the parentheses, so
stabParenthesesManyArguments is unneeded and can be replaced by
parenthesesArguments in stabParenthesesSignature.  This allows for
proper quoting of the unquote_splicing test cases from Elixir itself.  I
did not need to implement unwrap_splice.  I don't produce a wrapping
block for unquote_splicing as the elixir_parser.yrl expects.