%{ map | key => value }
Add updating to mapArguments.  Because the quoting is so different
between the previous mapArguments and updates, split the rule inside the
`{ }` to mapUpdateArguments and mapConstructionArguments.
mapUpdateARguments maps to just associationUpdate for now.
associationUpdate is `matched | matched => matched` with the `matched =>
matched` being matchedAssociationOperation so that associatonUpdate is
parsed like `|(current, =>(key, value))`.