Allow multiple EOL in endOfExpression, but only one endOfExpression
Elixir does not allow multiple `;`, in other words empty expressions,
but it does allow multiple newlines, so change endOfExpression to either
be `;` surrounded by any number of `EOL` or one or more `EOL` with not
`;`.  Then, to prevent multiple `;` in a row, change every
`endOfExpression+` to `endOfExpression` and change every
`endOfExpression*` to `endOfExpression?`.  Changes to testData were to
fix red tests for bug in anonymousFunction and for grouping to together
adjacent EOLs into one endOfExpression in other test datums.