Operator references
Support `<op>/<expression>` by parsing <op> as an IDENTIFIER instead of
its normal operator token so that `<op>/<expression>` will just be
parsed as a multiplication operation.  I attempted to write a new rule,
without changing the token the IDENTIFIER for the operator, but the
operator precedence is all messed up then.