Highlight Type Parameters
In @type, @typep, and @opaque, arguments to the type name are Type
Parameters that can be filled in later when the type is used.  The type
parameters appear in the definition of the type.  For @spec and
@callback, since the argument slots are already used for just the types
of the parameters, the type parameters are supplied as keyword keys in a
when clause.