Allow Infix#operator to work on operations with errors
Instead of asserting a set number of children, which is only true for
operations without errors in either operand (and these errors can occur
transiently when typing), just scan the children for the first one that
is an instanceof Operator.  This may fail if the right operand's error
format contains another Operator, but I wasn't able to generate that
format with manually testing.