Fix color constant deprecations
All of CodeInsightColors is deprecated, so all constants from there had
to be replaced.  Unfortunately, the recommended replacements don't have
the same color as the original, so I used different
DefaultLanguageHighlighterColors constants for some.

MODULE_ATTRIBUTE is now based on
DefaultLanguageHighlighterColors.CONSTANT (which is purplish in Darcula)
instead of the recommended METADATA, which is yellow.  Although module
attributes don't have to be constant since they can be set to
accumulate, often they are used as constants and not really as metadata,
since they are just data then.  All the `metadata` uses of module
attributes have a separate color.

SPECIFICATION is now based on
DefaultLanguageHighlighterColors.FUNCTION_DECLARATION, which maintains
the golden color that CodeInsightColors.METHOD_DECLARATION_ATTRIBUTES
had.

TYPE is now based on
DefaultLanguageHighlighterColors.METADATA, which is bright yellow unlike
CodeInsightColors.ANNOTATION_ATTRIBUTE_NAME_ATTRIBUTES, which was a
bright white.

TYPE_PARAMETER is now based on
DefaultLanguageHighlighterColors.PARAMETER, which unfortunately has no
attributes associated with it, but the constant name was too good a fit
not to use, so if you want the old color, you'll need to customize it
yourself.