Build jps-builder using only Java 1.6 compatible .classes
Fixes #290
Fixes #293
Fixes #295
Fixes #302
Fixes #313

In IntelliJ 14.1, all of openapi.jar targets Java 1.6 (with
MAJOR.MINOR 50.0), but in IntelliJ 2016.1, some of openapi.jar targets
only Java 1.8 (with MAJOR.MINOR 52.0), since jps-builders require parts
of openapi.jars and must target Java 1.6 even for IntelliJ 2016.1, the
52.0 .classes needed to be ported into org.elixir_lang.jps.builder, so
that the 52.0 version in openapi.jar wouldn't be attempted to be loaded.

This ended up being 5 classes:

* ExecutionException
* GeneralCommandLine
* ParametersList
* ParamsGroup
* ProcessNotCreatedException

Only GeneralCommandLine was used directly, all others are dependencies
of it.