canonicalName
canonicalName borrows from the idea of PsiReference#canonicalText: an
element can have both a Name (from getName), which is the literal name
in the code, which can be renamed, and a Canonical Name, which is the
name to refer to the element without need for imports or aliases.  For
this change, defimpl, defmodule, and defprotocol will show their full
module Alias for their Canonical name.

This change addresses the use case of Go To Declaration that should
resolved to a nested defmodule.