Add @doc template
A lot of ppl use the doc template after already typing @, but the `doc`
template starts with `@`, so it ends up inserting `@@doc ...`.  The
`@doc` is the same code, but since the name starts with `@`, it doesn't
insert a second `@`.  Unfortunately, the template search code doesn't
prompt when just typing `@`, so you end up having to type `@doc` before
only one template is selected.  The `@doc` template will show in the
lookup as soon as `@d` is typed, but you have to select it from the list
then before tabbing to accept.