Ship copy of GeneratedParserUtilBase
The generated ElixirParser uses the GeneratedParserUtilBase from
com.intellij.lang.parser, but since that version is a synced copy, it is
a snapshot of GeneratedParserUtilBase from the version of GrammarKit
that was current when the IDE was released.  To ensure the generated
parser works on all the IDEs, copy GeneratedParserUtilBase from
org.intellij.grammar.parser into org.elixir_lang.grammar.parser and then
use that version in ElixirParser.  This ensures neither the IDE's
version nor the version in any installed GrammarKit plugin is used.

NOTE: This copy needs to be updated if the build machine updates its
GrammarKit plugin.

So, I'm fixing errors from intellij-community having a sync-ed copy of
GeneratedParserUtilBase, by making my own synced copy...