Don't skip Call's parent when checking types in isParameter
isParameter(Call) called isParameter its parent if it wasn't a call
defintion clause, delegation or macro, but isParameter(PsiElement)
immediately calls getParent() and does all the instanceof tests on the
parent.  So, instead of isParameter(Call) calling
isParameter(PsiElement) on its parent, it should just call it on itself,
this way the check for ElixirInterpolation will not be skipped and
there's no need to handle ElixirInterpolatedString.