Add Mix ExUnit run configuration
This adds ExUnit test integration using the `mix test` command. It supports creating a run configuration manually or from a PSI element, meaning you can use the usual shortcut to run the test at the cursor's position, or run individual test files or directories. You can even right click on a particular test result to re-run just that test.

The latest elixir release (1.3.4) does not support specifying a custom test formatter on the CLI, but I submitted a change to enable that and it's been merged into master, so it should be released at some point. Until it is, I've bundled a custom mix task `TestWithFormatter`. Based on the Mix version in the "External Tools" settings, we guess whether this custom task is necessary or not.

I think the branch currently works, but it could use improvements in a few areas:
- Needs tests
- The RunConfigurationProducer will accept any Elixir file or directory. I'm not sure how best to limit it to test files and directories.
- Needs documentation

I think this is a really useful addition to the IDE, but I could potentially use some help (or at least feedback) to get it cleaned up a bit before merge. Feedback much appreciated.