Unresolved modular not treated as unqualified
maybeQualifiedCallToModular returned null BOTH if the call was
unqualified OR if the call was qualified, but its modular could not be
resolved, so qualified calls to .beam-only modules, like `File.read!`
returned `null` because `File` could not be resolved to a modular.
Remove `maybeqQualifiedToModular` and call `qualifiedToModular` when
`myElement` is qualified.  If the modular is `null`, then return an
empty ResolveResult[] instead of looking for unqualified matches.