Fix macro arity in decompiled code
Erlang version of Elixir macros have +1 arity because the Erlang
function takes the Caller as the first argument before taking the
arguments in the defmacro call.  This means when generating the
decompiled file -1 needs to be subjected from the arity.  Additionally,
with the correct arity, it is more obvious that some of the functions
names for #575 were invalid because they were operators, so for
operators place them infix in the defmacro call.