Remove call definition clauses completion for bare words
Completion of bare words as call definition clauses (functions or
macros) had a detrimental impact on typing feedback, which became
excessive once the index of call definition clauses was expanded by the
decompilation of the Elixir standard library `.beam`s, so disable it.
If bare-words completion is restored.  It will either (1) need to not
use the `Reference#getVariants()` API because it generates too many
objects that need to be thrown away or (2) need to only complete call
definition clauses that are provably in-scope from imports or other
macros.