Check if at end of file before checking for call opener
CallDefinitionClause InsertHandler always assumed there was a character
after the insert location, but when inserting at the very end of the
file, there isn't, which caused an IndexOutOfBoundsException.  Only
check if the following character is not a space, (, or [ when the
document is longer than the tail offset to prevent the
IndexOutOfBoundsException.  If the insertion is at the end of the file,
then the () will always be inserted.