Delegate ChildAttributes to previous child when available
Fixes #770

Block.getChildAttributes is used when Enter is pressed to determine the
indentation and alignment.  Using the default implementation, newlines
in stabs look overly indented, but there is a constant,
`DELEGATE_TO_PREV_CHILD` that can be used to just use the last child's
indent as long as there is one, which appears to work well for stabs (do
block bodies, etc).