Apply stabChildrenIndent rules to DO childAttributes
Fixes #793

When `do ... end` template is inserted, it did not have a previous child
whose attributes to use, so it used the default, which left the cursor
unindented.  To get the cursor indented by default, When the ElementType
is DO, apply the stabChildrenIndent rules: indent normal and determine
whether to indent relative to direct parent using code style setting.