Internal Erlang SDK for Elixir SDK
Elixir SDK will now have an Internal Erlang SDK, which can be supplied
one of two ways:

1. If `intellij-erlang` is installed: its Erlang SDK can be used
2. Otherwise: the Erlang SDK for Elixir SDK from this plugin can be used

Either Internal Erlang SDK's home path is used to locate the `erl` (or
`erl.exe` on Windows) executable and it is used to run Elixir; bypassing
the `elixir` (or `elixir.bat` on Windows) script.  Instead, as was done
in #789, the ebin directories for the Elixir SDK are passed as `-pa`
options to `erl`.

The Elixir SDK's configuration becomes more important: the classpath
entries are now the list of `ebin` directories to pass to `erl` using
`-pa` instead of scanning for ebins in the homepath, so additional code
paths can be added to the Classpaths configuration and they will be used
when running `mix` and `mix test`.  The classpaths will be initialized
to the ebin directories under the home path.