Decompress .beam if needed before decompiling
When an `.erl` file has `-compile([compressed])`, it will be compiled
normally, and then gzipped, so that the .beam will not contain the BEAM
magic number `FOR1`, but instead the gzip magic number 0x1f 0x8b.  So,
to properly handle normal and compressed, the decompiler needs to
look-ahead 2-bytes and check the gzip magic number.  If the gzip magic
number is detected, the InputStream will be passed through
GZIPInputStream before the the normal decoding process.