Put double quotes around macro names that have # in them
When decompiling function and macro names from Erlang module, its
possible for the name to contain `#` and they would be interpreted as
comments.  Accepting those names in Unquoted wouldn't escape the `#`
from being created as a comment, so in addition to `unquote`ing the
name, surround it in double quotes.