When setting up paths for Elixir SDK, setup and copy Erlang SDK paths
AdditionalDataConfigurable only copied the the Internal Erlang SDK code
paths when the selection in the ComboBox was changed, which meant if the
there was only one Erlang SDK or the user never changed the selection,
the Code Paths were never copied.  To get the Code Paths copied when the
Elixir SDK is first created and setupSdkPaths interface method is
called
1. All paths are set directly on the Elixir SDK as before
2. The default Erlang SDK, which should be the one that was last created,
is configured in SdkAdditionalaData
3. The Code paths from the Erlang SDK are copied to the Elixir SDK
  1. If the the Erlang SDK is from intellij-elixir and does not have
ebin directories for its Class Paths, then the Class Paths is expanded
to Code Paths.