Use SmallIDE setting for SDK path for elixir path.
Fixes #853

Rewrite of `MixRunningStateUtil#setElixir` to support calling `erl`
directly in IntelliJ IDEA added a check that sdk type was the Elixir SDK
Type, but there was no `else` clauses to handle that there IS an SDK,
but its NOT Elixir, which is the common case on all Small IDEs (like
Rubymine).  To fill this gap, if the SDK is not Elixir and a Small IDE
is detected, then the SDK Home from
Preferences > Other Settings > Elixir External Tools > Elixir SDK > Path
will be used to get the absolute path to the `elixir` (or `elixir.bat`
executable and set it in the GeneralCommandLine.

Tested with Rubymine on C-S-D/alembic running `mix test`.