Wrap runWriteAction in Type in invokeAndWait
Application#runWriteAction MUST be run from EventDispatch thread.  This
is true when run from a Run Configuration runner, but when the Credo
annotator runs runWriteAction, it's not in an EventDispatch thread, so
followed
https://www.jetbrains.org/intellij/sdk/docs/basics/architectural_overview/general_threading_rules.html
and wrapped the runWriteAction in Application#iinvokeAndWait.

Tested by setting up C-S-D/alembic SDK under 6.0.0, then updated to
master with changes: no errors occurred and credo issues were annotated.