Remove expensive !stabOperatorPrefix
Remove `!stabOperationPrefix` before `expression`, since
`stabOperationPrefix` itself starts with `expression` and instead put a
negative look-ahead, `!(infixComma | stabInfixOperator)` after
`expression`.  This now works when it previously didn't because `COMMA`
was added to `expressionRecoverUntil`, which means both `COMMA` and
`STAB_OPERATOR` are now in both not eaten by `expression`'s error
recovery.

This drops Fn1 through Fn8 to ~176ms instead of the previous time of the
last optimization of 13 seconds for Fn8.  It started at 26 seconds and
now it appears that the time no longer grows with the number of Fn.