LevelPropertyPusher
When the SDK changes, record what release of Elixir is used indirectly,
using a Level enum that only cares about the major and minor version.
Changing the Level, will mark the open files and .beam files as
invalidated, so that if the SDK is changed, the stubs and code Level
will agree.