Global credo inspection
Instead of using the ExternalAnnotatorBatchInspection, which runs the
credo Annotator on each file in batch mode, run `mix credo` once for
each working directory.  Also, since the tooltip is not preserved when
converting the Annotation to a ProblemDescriptor, don't calculate the
explanation, which eliminates the extra `mix credo PATH:LINE[:COLUMN]`
for each issue, which makes the batch quite fast.