Don't allow usage of Erlang SDK in Small IDEs
Erlang SDK from intellij-erlang plugin is not meant to be created in
Small IDEs because it uses `JavadocOrderRootType.getInstance()`, which
will fail in Small IDEs.  Because Erlang SDK is never creatable, even if
intellij-erlang is installed, if in a Small IDE, always use the Erlang
SDK for Elixir SDK that comes from intellij-elixir.