Convert MFA to file in debugger stack
When the module, function, and arguments are available for each stack
frame, the module can be checked to see if it has a __templates__/0
function and if the functon name in in the __templates__() output.  If
it is, then the modules file should be the templates path and not the
module's compile source.  Because the debugger OTP app uses Erlang
Abstract Code, it only uses the line number and not the filename like it
would if it used the Code and Line chunks together.