Debugging for IEx Run Configurations
Rewrite how debugging works across all run configurations:

1. Java Node starts TCP port as before
2. Debugger Elixir node is started with and runs mix
intellij_elixir.debug
3. Debugged Elixir node is started

With (2) and (3) as separate nodes now, there needs to be a way for the
debugger to connect to the debugged node and for the interpreted modules
to be used in the debugged node.  For this coordination the follow is
used

1. BOTH the debugger and debugged Elixir node use a shared `-config
sys.config` that requires both nodes to be connected to proceed:
   ```erlang
   [{kernel
     [
       {sync_nodes_mandatory, ['debugged<UUID>@127.0.0.1', 'debugger<UUID>@127.0.0.1']},
       {sync_nodes_timeout, 10000}
     ]}
   ].
   ```
2. The debugged Elixir node requires
   `resources/debugged/lib/intellij_elixir/debugged.ex` using Erlang:
   1. `-eval application:ensure_all_started(elixir)`
   2. `-eval 'Elixir.Code':require_file(<<"path/to/debugged.ex">>)
3. `debugged.ex` registers `self()` as `IntelliJElixir.Debugged` and
enters a `receive` that waits for a `:continue` `GenServer.call` from
the debugger Elixir node.
4. The Java XDebugProcess gets to `sessionInitialized`.
  1. Java Node is started
  2. Elixir Debugger node is started
  3. Java Node sends `{:connect, debugged_name}` to the Elixir Debugger
node
5. The Elixir Debugger node received {:connect, debugged_name}
  1. `Node.connect(debugged_name)`, which should already be true because
of the `sys.config`, but is a check that it hasn't died.
  2. `GenServer.call({IntelliJElixir.Debugged, debugged_name},
:continue)`
6. The Elixir debugged node receives the `GenServer.call` and continues,
with the debugger properly attached and the `:int` module globally
interpreted from the debugger to the debugged node before the debugged
node runs any code because the `-eval` is sequential and blocking for
`erl`.

In addition to this improved startup, the debugger now supports a
`:stop` cast, so that the Java node can tell the Elixir debugger node to
stop itself, which ensures that the `beam.smp` for the debugger no
longer sticks around.  The problem was that the IntelliJ Process class
cannot kill `beam.smp` normally because of how it intercepts OS signals.