Show Elixir variable instead of Erlang variable
Whenever possible, so only the Elixir variable name instead of the
Erlang `V<elixirName>@<bindingIndex>` name.  When one Elixir variable is
rebound, group together the Previous Bindings into a XValueGroup that is
collapsed by default and show the current Erlang variable name and value
(since we can't show a value without a name when the XValueGroup is a
sibling).