Improve error logging for unexpected element class unaliased name d wn
Fixes #1204

`org.elixir_lang.reference.module.UnaliasedName.down(PsiElement)` only
expects `ElixirAccessExpression` or `QualifiableAlias`, but for
unexpected classes of elements, it was a simple Kotlin `TODO()`, which
didn't log anything useful for enhancements.  Replace the `TODO()` with
an `org.elixir_lang.errorreport.Logger.error`, so that the element's
class and content can be reported when this happens again.