Don't sync libraries on initComponent
For an unknown reason, when sync occurs at `initComponent` time in
`DepsWatcher` or `mix.Watcher`, the child directories of the project
basedDir aren't shown in the Project Pane until a change is applied in
Project Structure.

Turning off the syncing in `initComponent` means that people upgrading
to 10.0.0 will not get the automatic Libraries and Modules until they
next change those directories, when the VirtualFileListener will pick up
the change and sync.