Run library syncing in tasks
When importing projects, run sync as a modal task while when
VirtualFileSystem events occur, run sync as background tasks.  Either
style of task is cancellable and importantly, it allows to show what
is happening during the syncing.  This converts the unresponsive UI
beachball on macOS during import to an updating progress bar.